組み合わせ計算処理の切り出し
